Rename 'object' to 'envObject'
