+++ src/main/java/fr/utbm/info/vi51/worldswar/environment/Cell.java
@@ -3,20 +3,20 @@ package fr.utbm.info.vi51.worldswar.environment;
 import java.util.LinkedList;
 import java.util.List;
 
-import fr.utbm.info.vi51.worldswar.environment.object.EnvironmentObject;
+import fr.utbm.info.vi51.worldswar.environment.envobject.EnvironmentObject;
 
 /**
  * Stores the content of a tile in the environment grid.
  *
  */
 public class Cell {
-	private List<EnvironmentObject> objects;
+	private List<EnvironmentObject> envObjects;
 
 	public Cell() {
-		this.objects = new LinkedList<>();
+		this.envObjects = new LinkedList<>();
 	}
 
-	public List<EnvironmentObject> getObjects() {
-		return this.objects;
+	public List<EnvironmentObject> getEnvObjects() {
+		return this.envObjects;
 	}
 }
