+++ src/main/java/fr/utbm/info/vi51/worldswar/environment/perceivable/PerceivableAnt.java
@@ -2,7 +2,7 @@ package fr.utbm.info.vi51.worldswar.environment.perceivable;
 
 import fr.utbm.info.vi51.worldswar.environment.Caste;
 import fr.utbm.info.vi51.worldswar.environment.Colony;
-import fr.utbm.info.vi51.worldswar.environment.object.AntBody;
+import fr.utbm.info.vi51.worldswar.environment.envobject.AntBody;
 
 /**
  * {@link Perceivable} representation of an {@link AntBody}
@@ -18,7 +18,7 @@ public class PerceivableAnt extends Perceivable {
 
 	/**
 	 * @return
-	 * @see fr.utbm.info.vi51.worldswar.environment.object.AntBody#getFoodCarried()
+	 * @see fr.utbm.info.vi51.worldswar.environment.envobject.AntBody#getFoodCarried()
 	 */
 	public int getFoodCarried() {
 		return this.ant.getFoodCarried();
@@ -26,7 +26,7 @@ public class PerceivableAnt extends Perceivable {
 
 	/**
 	 * @return
-	 * @see fr.utbm.info.vi51.worldswar.environment.object.AntBody#getHealth()
+	 * @see fr.utbm.info.vi51.worldswar.environment.envobject.AntBody#getHealth()
 	 */
 	public int getHealth() {
 		return this.ant.getHealth();
@@ -34,7 +34,7 @@ public class PerceivableAnt extends Perceivable {
 
 	/**
 	 * @return
-	 * @see fr.utbm.info.vi51.worldswar.environment.object.AntBody#getCaste()
+	 * @see fr.utbm.info.vi51.worldswar.environment.envobject.AntBody#getCaste()
 	 */
 	public Caste getCaste() {
 		return this.ant.getCaste();
@@ -42,7 +42,7 @@ public class PerceivableAnt extends Perceivable {
 
 	/**
 	 * @return
-	 * @see fr.utbm.info.vi51.worldswar.environment.object.AntBody#getColony()
+	 * @see fr.utbm.info.vi51.worldswar.environment.envobject.AntBody#getColony()
 	 */
 	public Colony getColony() {
 		return this.ant.getColony();
