+++ src/main/java/fr/utbm/info/vi51/worldswar/environment/perceivable/PerceivableWall.java
@@ -1,6 +1,6 @@
 package fr.utbm.info.vi51.worldswar.environment.perceivable;
 
-import fr.utbm.info.vi51.worldswar.environment.object.Wall;
+import fr.utbm.info.vi51.worldswar.environment.envobject.Wall;
 
 /**
  * {@link Perceivable} representation of a {@link Wall}
